+++ rocprim/include/rocprim/intrinsics/warp_shuffle.hpp
@@ -254,15 +254,39 @@ template<class V>
 ROCPRIM_DEVICE ROCPRIM_INLINE
 V warp_swizzle_shuffle(V& v, const int mask, const int width = arch::wavefront::min_size())
 {
-    switch(mask)
+    using VV = typename std::remove_cv<V>::type;
+
+    VV result = v;
+
+    // __builtin_amdgcn_ds_swizzle instruction requires the offset to be a compile-time constant, so we
+    // cannot pass a runtime variable `mask` directly.
+    //
+    // According to the AMD ISA documentation, the offsets for the Bitwise XOR swizzle mode is encoded
+    // as `(XOR_MASK << 10) | 0x1F`. We can use this formula to get the offsets from the XOR masks during
+    // complile time.
+    //
+    // We use compile-time unrolling to generate branches for the corresponding `ds_swizzle_b32` instructions.
+    ::rocprim::detail::constexpr_for_lt<1, 32, 1>(
+        [&](auto i)
+        {
+            if(mask == i)
+            {
+                result = warp_swizzle<V, (i << 10) | 0x1F>(v);
+            }
+        });
+
+    // Fallback branch:
+    // `ds_swizzle_b32` is limited to 32-lane physical boundaries and supports a maximum 5-bit XOR mask
+    // (0 to 31). If the mask is >= 32 (e.g., 32 or 63 in a Wave64), it physically cannot be performed
+    // via swizzle.
+    //
+    // We fall back to the generic `warp_shuffle_xor`, which degrades to `ds_bpermute_b32` (LDS crossbar
+    // routing), it's slower but necessary for cross-half-wave operations.
+    if(mask >= 32)
     {
-        case 1: return warp_swizzle<V, 0x041F>(v);
-        case 2: return warp_swizzle<V, 0x081F>(v);
-        case 4: return warp_swizzle<V, 0x101F>(v);
-        case 8: return warp_swizzle<V, 0x201F>(v);
-        case 16: return warp_swizzle<V, 0x401F>(v);
-        default: return warp_shuffle_xor(v, mask, width);
+        return warp_shuffle_xor(v, mask, width);
     }
+    return result;
 }
 
 } // namespace detail