+++ CHANGELOG.md
@@ -29,6 +29,10 @@ Full documentation for rocPRIM is available at [https://rocm.docs.amd.com/projec
 * Reduced build times for unit tests.
 * Memory usage in unit tests.
 
+### Changed
+
+* Changed the `bitonic_sort` algorithm in `warp_sort_shuffle` to use forward-only comparison for better performance. `block_sort_bitonic` is also changed to use forward-only comparison, to align the sorting with `warp_sort`.
+
 ### Resolved issues
 
 * Fixed a silent overflow in `rocprim::device_segmented_reduce` where it could exceed the maximum number of HIP threads, resulting in missing output.